Complete Red-Green User Story 1 (US1)
∆ Delete User Story 2 (US2). Looking at requirements for US2 and it is
clear that testing for interaction with multiple rovers will include
the same expects as testing for a singular rover. This is due to the
design decision to inject Mars Rovers into the Controller class. US2
has now become vacuous.

∆ Refactor feature test suit, #Controller.online? now expects a boolean
to be returned as opposed to the simpler #respond_to expect. This is due
to the requirement that we test that the object passed in is online, not
that the Controller object responds to a method #online?

∆ Refactor Controller class, #online? now returns the value of its
argument when passed a method #online. We are now returning the result of
calling #online on any object that is passed to Controllers #online?
method.

∆ Add #online to MarsRover class, this is currently hard coded to
return true. MarsRover class will now always send a message containing
'true' when another class (Controller in this case), asks it its state.