Work on Red stage of User Story 2 (US2)
∆ Add feature test for US2. This test that the controller object can
implement the #land behaviour and in doing so the plateau object shall
receive a message to let it know an object has landed on it.

∆ Add unit test for Controller #land. This test that the controller can
send a message instructiong an object to land.

∆ Add unit test for Plateau #curently_landed_vehicles. This tests that
on receiving the correct message the Plateau object can store objects
that respond to #land.

∆ Add method #land to Controller. This currently takes two arguments
(vehicle and location). This method is the basis of sending the land
method, where the first argument is the object it will try to land
and the second argument is the object it will try to land on.

∆ Add #currently_landed_vehicles to Plateau. This is where a data
structure shall be used to store to the state of a Plateau object
the objects are landed on a plateau.