Add #landed to MarsRover class
∆ Hardcode state of MarsRover to return 'true' for landed method. This
method will send the mesage to the controller that the MarsRover has
successfully landed at the location passed into the #land method in
Controller as its second argument. This passes the unit test in
'controller_spec' for 'can land an object'.

∆ Add few additional changes as currently got errors for different
features. This is going a little off TDD, be aware, now trying to bring
it back. Changes are as follows:

∆ Refactor 'user_story_spec' to expect a different method in feature
test for US2. Now expects #number_of_landed_rovers as an integer. This
is because I thought a complex data structure as was being expect with
previous method #currently_landed_vehicles was actually answering a
different problem. The current problem is 'is it possible to land
something', while the problem answered by #currently_landed_vehicles
is 'what is landed at the moment'.

∆ Refactor 'plateau_spec' so unit tests expected the same as feature
tests. Same changes and reasons as the point above.

∆ Add state to a Plateau class on initialize that stores the
NASA controller can send messages to and recive messages from that
inform them of the amount of rovers currently on the Plateau.