Complete Red-Green stages of User Story 2 (US2)
∆ Add argument to #landed called in 'controller_spec'. The
MarsRover will now take an argument of a location for where it has
been landed. This will be inherited from Controller #land.

∆ Refactor argument in 'plateau_spec', removed subject and used plateau
variable set 3 lines above it. This adds a location to where the
controller will land a MarsRover. A little unsure why subject didn't
work, definately needs some refactoring and looking into further.

∆ Add #landed(location) to MarsRover. This writes to the Plateau class
adding an additional MarsRover to the Plateau.

∆ Refactor @landed to initialize as 'false' and use #landed(location)
to set @landed to true when method is called. This is so the state of
a MarsRover starts as not landed and can now be landed by the
controller. Concern over #landed(location) as it no longer adhears
to SRP.

∆ Refactor Plateau so :number_of_landed_rovers is an attr_accessor.
access to this method.

∆ Refactor #land in Controller to now call #landed(location) in
MarsRover. This is so the controller can land the mars rover and in
turn write to the Plateau as explained above.