Refactor Controller class
∆ Refactor argument in #online? from 'mars_rover' to 'vehicle'. This
is ab advantage of dependency injection, the Controller class can check
the state of any object that responds to #online?. Therefore, in future
advances of the program, another designer may want the controller to
check state of other objects such as landing pods or scientific
equipment. The method as it is now set up provides easily changable
code to facilitate this.