Complete Red-Green stages of User Story 4 (US4)
∆ Add unit test 'can return its current position'. This expects
MarsRover to return two coordinates and a heading. Meets criteria set in
US4

∆ Refactor #vehicle_position in Controller to return the #current_position
of the vehicle passed in as an argument. This will send the message to
the MarsRover class with the information required to passed the feature
test.

∆ Add #current_position to MarsRover. Returns a string including the two
coordinates and the heading.

∆ Set up writable methods #x_coordinate, #y_coordinate and #heading.
These will each have the single responsibility of managing the elements
of the MarsRovers position that they refer to.