Complete Red-Green stages for coordinate unit test
∆ Refactor 'controller_spec' test 'can check the maximum boundries of
the grid'. This now checks for a behaviour rather than a response
with no behaviour.

∆ Add unit test to 'plateau_spec' to check for a response when
questioned about the size of the current grid laid over the plateau.

∆ Refactor Controller #check_location_bounds to return the response
from the Plateau when questioned about the size of the current
overlaid grid.

∆ Add #grid_size which returns a string that informs of the maximum
x and y coordinates. On initialization, no grid has been laid over
the plateau so the expect is therefore expecting nil. A future
feature not asked for in the challenge could be an error raised
when no grid has been overlaid.

∆ Add state of #x_axis_max and #y_axis_max to plateau. These
seperate the responsibility for each part of the setting of coords.

∆ I am considering whether the grid is really the responsibility
of the plateau. Should there be a seperate Grid class? Am am
starting to think so.